Refactored json read to not extends from FilePartition
