+++ connector/src/main/scala/com/vertica/spark/datasource/partitions/file/VerticaFilePartition.scala
@@ -24,10 +24,11 @@ import org.apache.spark.sql.execution.datasources.{FilePartition, PartitionedFil
  * */
 class VerticaFilePartition(override val index: Int,
                            override val files: Array[PartitionedFile],
+                           val filesIdentity: Array[PartitionedFileIdentity],
                            val partitioningRecords: Map[String, Int])
   extends FilePartition(index, files) with Cleanup {
 
-  override def getPortions: Seq[Identifiable] = this.files.map(_.asInstanceOf[Identifiable])
+  override def getPortions: Seq[Identifiable] = this.filesIdentity
 
   override def getPartitioningRecord: Map[String, Int] = this.partitioningRecords
 }
