+++ connector/src/main/scala/com/vertica/spark/datasource/wrappers/VerticaScanWrapper.scala
@@ -14,10 +14,12 @@
 package com.vertica.spark.datasource.wrappers
 
 import com.vertica.spark.config.ReadConfig
-import com.vertica.spark.datasource.partitions.file.{VerticaFilePartition, VerticaFilePortion}
+import com.vertica.spark.datasource.partitions.file.{PartitionedFileIdentity, VerticaFilePartition}
 import org.apache.spark.sql.connector.read.{Batch, InputPartition, PartitionReaderFactory, Scan}
 import org.apache.spark.sql.execution.datasources.{FilePartition, PartitionedFile}
 import org.apache.spark.sql.types.StructType
+
+import scala.collection.mutable.ArrayBuffer
 
 /**
  * Wraps a [[Scan]] so that it will create a [[PartitionReaderWrapperFactory]]
@@ -33,21 +35,23 @@ class VerticaScanWrapper(val scan: Scan, val config: ReadConfig) extends Scan wi
    * */
   override def planInputPartitions(): Array[InputPartition] = {
     val partitioningRecords = scala.collection.mutable.Map[String, Int]()
+    val fileIdentities = new ArrayBuffer[PartitionedFileIdentity]()
 
     /**
      * Make a file portion and record it.
      * */
-    def makeFilePortion(file: PartitionedFile) = {
+    def makeFilePortion(file: PartitionedFile): PartitionedFile = {
       val key = file.filePath
       val count = partitioningRecords.getOrElse(key, 0)
       partitioningRecords.put(key, count + 1)
-      VerticaFilePortion(file, count)
+      fileIdentities.append(PartitionedFileIdentity(file.filePath, file.start))
+      file
     }
 
     scan.toBatch.planInputPartitions()
       .map(partition => partition.asInstanceOf[FilePartition])
       .map(filePartition => filePartition.copy(files = filePartition.files.map(makeFilePortion)))
-      .map(partition => new VerticaFilePartition(partition.index, partition.files, partitioningRecords.toMap))
+      .map(partition => new VerticaFilePartition(partition.index, partition.files, fileIdentities.toArray, partitioningRecords.toMap))
   }
 
   override def createReaderFactory(): PartitionReaderFactory = {
