+++ connector/src/test/scala/com/vertica/spark/datasource/wrappers/VerticaScanWrapperTest.scala
@@ -1,8 +1,7 @@
 package com.vertica.spark.datasource.wrappers
 
 import com.vertica.spark.common.TestObjects
-import com.vertica.spark.datasource.partitions.file.{VerticaFilePartition, VerticaFilePortion}
-import com.vertica.spark.util.schema.SchemaToolsTests.mock
+import com.vertica.spark.datasource.partitions.file.VerticaFilePartition
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.connector.read.{Batch, InputPartition, PartitionReaderFactory, Scan}
 import org.apache.spark.sql.execution.datasources.{FilePartition, PartitionedFile}
@@ -49,7 +48,6 @@ class VerticaScanWrapperTest extends AnyFlatSpec with BeforeAndAfterAll with Moc
 
     val verticaFilePartition = partitions.head.asInstanceOf[VerticaFilePartition]
     assert(verticaFilePartition.index == 1)
-    assert(verticaFilePartition.files.head.isInstanceOf[VerticaFilePortion])
 
     val fileCounts = verticaFilePartition.partitioningRecords.keySet.toList.length
     assert(fileCounts == 4)
